first pass at creating context for nav
